login-service/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
passlib[bcrpyt]
httpx
asyncpg